feat(LruCacheClient): implement LruCacheClient
